refactor: (#5) OpenFeign 관련 패키지 구조 변경
OpenFeign 관련 클래스들을 `global/utils`에서 `infrastructure`
패키지로 이동

- OpenFeign 기능이 외부 시스템과의 연동을 담당하는
  인프라 계층에 속한다고 판단하여 패키지 이동
- 향후 인프라 관련 코드들을 `infrastructure` 패키지
  하위에 추가 관리 예정
