+++ src/main/kotlin/org/example/root_be/infrastructure/openfeign/FeignClientErrorDecoder.kt
@@ -1,9 +1,9 @@
-package org.example.root_be.global.utils.openfeign
+package org.example.root_be.infrastructure.openfeign
 
 import feign.Response
 import feign.codec.ErrorDecoder
-import org.example.root_be.global.utils.openfeign.client.exception.FeignBadRequestException
-import org.example.root_be.global.utils.openfeign.client.exception.FeignServerException
+import org.example.root_be.infrastructure.openfeign.client.exception.FeignBadRequestException
+import org.example.root_be.infrastructure.openfeign.client.exception.FeignServerException
 import org.springframework.stereotype.Component
 
 @Component
