+++ src/main/kotlin/org/example/root_be/infrastructure/openfeign/client/DsmAuthClient.kt
@@ -1,8 +1,8 @@
-package org.example.root_be.global.utils.openfeign.client
+package org.example.root_be.infrastructure.openfeign.client
 
-import org.example.root_be.global.utils.openfeign.client.dto.request.UserInfoRequest
+import org.example.root_be.infrastructure.openfeign.client.dto.request.UserInfoRequest
 import org.springframework.cloud.openfeign.FeignClient;
-import org.example.root_be.global.utils.openfeign.client.dto.response.UserInfoResponse
+import org.example.root_be.infrastructure.openfeign.client.dto.response.UserInfoResponse
 import org.springframework.web.bind.annotation.PostMapping
 
 @FeignClient(name = "dsm-auth", url = "\${dsm.auth.url}")
