+++ src/main/kotlin/org/example/root_be/infrastructure/openfeign/client/exception/FeignServerException.kt
@@ -1,4 +1,4 @@
-package org.example.root_be.global.utils.openfeign.client.exception
+package org.example.root_be.infrastructure.openfeign.client.exception
 
 import org.example.root_be.global.err.exception.CustomException
 import org.example.root_be.global.err.exception.ErrorCode
